backend/curator/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.112.0
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.4.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
requests==2.32.3
scikit-learn==1.5.2
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
